# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Server",
    "title": "exchange API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "Ivan Anidi",
      "url": "https://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:4000",
  "basePath": "/",
  "paths": {
...
    "/auth/phone/request": {
      "post": {
        "description": "PhoneRequestCode",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Public", "Auth"],
        "summary": "запрос кода для входа по смс",
        "operationId": "Phone-Auth",
        "parameters": [
          {
            "type": "string",
            "description": "Телефон пользователя в формате 79999999999",
            "name": "Phone",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/public.Settings"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/public.Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "public.Error": {
      "type": "object",
      "properties": {
        "Error": {
          "type": "string",
          "example": "ERROR_MESSAGE"
        },
        "Status": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "public.Settings": {
      "type": "object",
      "properties": {
        "AdminEmail": {
          "description": "Email адрес для связи с администрацией",
          "type": "string"
        },
        "AdminTelegram": {
          "description": "Telegram аккаунт для связи с администрацией",
          "type": "string"
        },
        "BookInAdvanceDaysRestriction": {
          "description": "За сколько дней заблаговременно можно забронировать переговорную комнату",
          "type": "integer"
        },
        "MaxSlotsPerBooking": {
          "description": "Максимальное количество слотов на одно бронирование",
          "type": "integer"
        },
        "PushChatNewMessage": {
          "description": "Текст нового сообщения в чате",
          "type": "string"
        },
        "SlotDuration": {
          "description": "Размер слота бронирования в минутах",
          "type": "integer"
        },
        "TimeForBookingConfirmation": {
          "description": "Время на подтверждение переговорной комнаты в минутах",
          "type": "integer"
        },
        "URLAbout": {
          "description": "Ссылка о проекте",
          "type": "string"
        },
        "URLLicense": {
          "description": "Ссылка на лицензии",
          "type": "string"
        },
        "URLPrivacy": {
          "description": "Ссылка на политику конфиденциальности",
          "type": "string"
        },
        "URLServices": {
          "description": "Ссылка на услуги",
          "type": "string"
        },
        "URLTOS": {
          "description": "Ссылка на условия использования",
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}